requirements.txt
pandas==2.0.1
mysql-connector-python==8.0.33
openpyxl==3.1.2
pymysql==1.0.3
sqlalchemy
pymysql